adjusts crud sample to new implementation of crud support

+++ leon-samples/angular/curd/src/main/webapp/index.server.coffee
@@ -2,15 +2,20 @@
 db = -> leon.mongo.addresses
 
 @indexService =
-    list: ->
-        db().find().sort({name:1, mail:-1}).toArray()
+	list: ->
+		db().find().sort({name:1, mail:-1}).toArray()
 
-    get: (id) ->
-        db().findByOId(id)
+	get: (id) ->
+		db().findByOId(id)
 
-    save: (data) ->
-        db().save(data)
+	save: (data) ->
+		db().save(data)
 
-    delete: (id) ->
-        data = db().findByOId(id)
-        db().remove(data)
+	delete: (id) ->
+		data = db().findByOId(id)
+
+		if data?
+			db().remove(data)
+			true
+		else
+			false
